+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- Added a `.filled` operation to `Mosaic` and `ImageStacks` that allows one to fill the masked values of a masked array with a given value
+
 ## v1.3.3 - 11/13/2024
 
 ### Added